pyproject.toml
[project]
name = "dds-manager"
version = "0.1.0"
description = "Django app for cash flow management and transaction tracking"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "django>=5.2.4",
    "django-smart-selects>=1.7.2",
]

[dependency-groups]
dev = [
    "black>=25.1.0",
    "django-debug-toolbar>=6.0.0",
    "ruff>=0.12.5",
]

# ruff settings
[tool.ruff]
required-version = ">=0.12.5"

src = [
    "dds_manager",
]

# Exclude a variety of commonly ignored directories.
exclude = [
    "__pycache__",
    "__pypackages__",
    "_build",
    ".bzr",
    ".direnv",
    ".eggs",
    ".git",
    ".git-rewrite",
    ".hg",
    ".ipynb_checkpoints",
    ".mypy_cache",
    ".nox",
    ".pants.d",
    ".pyenv",
    ".pytest_cache",
    ".pytype",
    ".ruff_cache",
    ".svn",
    ".tox",
    ".venv",
    ".vscode",
    "build",
    "buck-out",
    "dist",
    "node_modules",
    "site-packages",
    "venv",
]

extend-exclude = [
    "migrations",
    "dds_manager/manage.py",
]

# Same as Black.
line-length = 88
indent-width = 4

# Assume Python 3.13
target-version = "py313"

[tool.ruff.lint]
select = [
    "FAST", # FastAPI (FAST)
    "YTT", # flake8-2020 (YTT)
    "ANN", # flake8-annotations (ANN)
    # "ASYNC", # flake8-async (ASYNC)
    "S", # flake8-bandit (S)
    "BLE", # flake8-blind-except (BLE)
    "FBT", # flake8-boolean-trap (FBT)
    "B", # flake8-bugbear (B)
    "COM", # flake8-commas (COM)
    "C4", # flake8-comprehensions (C4)
    "DTZ", # flake8-datetimez (DTZ)
    "T10", # flake8-debugger (T10)
    "DJ", # flake8-django (DJ)
    "EM", # flake8-errmsg (EM)
    "ISC", # flake8-implicit-str-concat (ISC)
    "ICN", # flake8-import-conventions (ICN)
    "LOG", # flake8-logging (LOG)
    "G", # flake8-logging-format (G)
    "INP", # flake8-no-pep420 (INP)
    "PIE", # flake8-pie (PIE)
    "T20", # flake8-print (T20)
    "PYI", # flake8-pyi (PYI)
    # "PT", # flake8-pytest-style (PT)
    "Q", # flake8-quotes (Q)
    "RSE", # flake8-raise (RSE)
    "RET", # flake8-return (RET)
    "SLF", # flake8-self (SLF)
    "SIM", # flake8-simplify (SIM)
    "SLOT", # flake8-slots (SLOT)
    "TC", # flake8-type-checking (TC)
    "ARG", # flake8-unused-arguments (ARG)
    "PTH", # flake8-use-pathlib (PTH)
    "FLY", # flynt (FLY)
    "I", # isort (I)
    "C90", # mccabe (C90)
    "N", # pep8-naming (N)
    "PERF", # Perflint (PERF)
    "E", # pycodestyle Error (E)
    "W", # pycodestyle Warning (W)
    # "DOC", # pydoclint (DOC)
    # "D", # pydocstyle (D)
    "F", # Pyflakes (F)
    "PGH", # pygrep-hooks (PGH)
    "PLC", # Pylint Convention (PLC)
    "PLE", # Pylint Error (PLE)
    "PLR", # Pylint Refactor (PLR)
    "PLW", # Pylint Warning (PLW)
    "UP", # pyupgrade (UP)
    "FURB", # refurb (FURB)
    "RUF", # Ruff-specific rules (RUF)
    "TRY", # tryceratops (TRY)
]
ignore = [
    "ANN401", # Checks that function arguments are annotated with a more specific type than Any.
    "E501", # Checks for lines that exceed the specified maximum character length.
    "RUF001", # Checks for ambiguous Unicode characters in strings.
    "RUF012", # Checks for mutable default values in class attributes.
]

# Allow fix for all enabled rules (when `--fix`) is provided.
fixable = ["ALL"]
unfixable = []
